# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: EventHub API
  version: "1.0.0"
  description: "Documentation \"EventHub\" API v1.0"
servers:
  - url: http://localhost:8080
...
paths:
  /users/v1/api:
    post:
      tags:
        - user
      summary: Создать пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDto'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь с таким username уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - user
      summary: Получить список пользователей
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserDtoResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/v1/api/{id}:
    get:
      tags:
        - user
      summary: Получить пользователя по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Некорректный идентификатор
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - user
      summary: Обновить пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDto'
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь с таким username уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - user
      summary: Удалить пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Пользователь удалён
        '400':
          description: Некорректный идентификатор
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/v1/api/metadata:
    get:
      tags:
        - user-metadata
      summary: Получить список метаданных пользователей
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Список метаданных пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserMetadataDtoResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/v1/api/{id}/metadata:
    post:
      tags:
        - user-metadata
      summary: Создать метаданные пользователя для указанного userId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserMetadataCreateDto'
      responses:
        '201':
          description: Метаданные пользователя успешно созданы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMetadataDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - user-metadata
      summary: Получить метаданные пользователя по userId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Метаданные пользователя найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMetadataDto'
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Метаданные пользователя не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - user-metadata
      summary: Обновить метаданные пользователя по userId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserMetadataUpdateDto'
      responses:
        '200':
          description: Метаданные пользователя обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMetadataDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Метаданные пользователя не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - user-metadata
      summary: Удалить метаданные пользователя по userId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Метаданные пользователя удалены
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Метаданные пользователя не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    UserCreateDto:
      type: object
      required: [ username ]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Имя пользователя
          example: "ivan.petrov"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "ivan.petrov@example.com"
    UserUpdateDto:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          description: Новое имя пользователя
          example: "petr.ivanov"
        email:
          type: string
          format: email
          description: Новый email пользователя
          example: "petr.ivanov@example.com"
    UserMetadataCreateDto:
      type: object
      properties:
        firstName:
          type: string
          description: Имя
          example: "Иван"
        lastName:
          type: string
          description: Фамилия
          example: "Петров"
        bio:
          type: string
          description: Краткая биография
          example: "Java-разработчик, спикер митапов."
        phone:
          type: string
          maxLength: 15
          description: Телефон
          example: "+7-999-123-45-67"
    UserMetadataUpdateDto:
      type: object
      properties:
        firstName:
          type: string
          description: Новое имя
          example: "Пётр"
        lastName:
          type: string
          description: Новая фамилия
          example: "Иванов"
        bio:
          type: string
          description: Новая биография
          example: "Backend-разработчик."
        phone:
          type: string
          maxLength: 15
          description: Новый телефон
          example: "+7-888-654-32-10"
...
    UserShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор пользователя
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        username:
          type: string
          description: Имя пользователя
          example: "ivan.petrov"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "ivan.petrov@example.com"
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор пользователя
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        username:
          type: string
          description: Имя пользователя
          example: "ivan.petrov"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "ivan.petrov@example.com"
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания пользователя
          example: "2025-12-01T10:00:00+03:00"
    UserMetadataDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор метаданных пользователя
          example: "d1e2f3a4-5678-1234-abcd-9876543210ab"
        user:
          $ref: '#/components/schemas/UserShortDto'
        firstName:
          type: string
          description: Имя
          example: "Иван"
        lastName:
          type: string
          description: Фамилия
          example: "Петров"
        bio:
          type: string
          description: Краткая биография
          example: "Java-разработчик, спикер митапов."
        phone:
          type: string
          maxLength: 15
          description: Телефон
          example: "+7-999-123-45-67"
        createdAt:
          type: string
          format: date-time
          description: Дата создания метаданных
          example: "2025-12-01T10:00:00+03:00"
        updatedAt:
          type: string
          format: date-time
          description: Дата обновления метаданных
          example: "2025-12-02T12:00:00+03:00"
...
    ErrorResponse:
      type: object
      properties:
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: "/events"
        httpMethod:
          type: string
          description: HTTP метод запроса
          example: "POST"
        statusCode:
          type: integer
          description: Код ошибки
          example: 400
        errorReason:
          type: string
          description: Краткое описание причины ошибки
          example: "Bad Request"
        message:
          type: string
          description: Описание ошибки
          example: "Некорректные данные запроса"
...
    PagedUserDtoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    PagedUserMetadataDtoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserMetadataDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...